requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.14.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
docutils==0.20.1
drf-nested-routers==0.93.4
idna==3.6
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Pygments==2.17.2
pytz==2023.3.post1
requests==2.31.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
